Everything is done. Need to test

+++ LibraryApi.Write/Services/WriteRepository/WriteDbRepository.cs
@@ -14,6 +14,9 @@ public class WriteDbRepository: IWriteRepository<Book>
 	
 	public Book Create(Book book)
 	{
+		if (_db.Books.Any(b => b.ISBN == book.ISBN))
+			throw new ArgumentException($"ISBN {book.ISBN} is already occupied");
+		
 		var ee = _db.Books.Add(book);
 		_db.SaveChanges();
 
@@ -22,17 +25,29 @@ public class WriteDbRepository: IWriteRepository<Book>
 
 	public Book Update(Book book)
 	{
+		if (!_db.Books.Any(b=>b.Id == book.Id))
+			throw new KeyNotFoundException($"There is no book with id {book.Id}");
+		if (_db.Books.Any(b => b.ISBN == book.ISBN && b.Id != book.Id))
+			throw new ArgumentException($"ISBN {book.ISBN} is already occupied");
+		
+		Console.WriteLine($"In updatte. Id {book.Id}, isbn {book.ISBN}");
 		var ee = _db.Books.Update(book);
+		var res = ee.Entity;
+		Console.WriteLine("After update");
 		_db.SaveChanges();
 
-		return ee.Entity;
+		return res;
 	}
 
 	public Book Delete(int id)
 	{
+		if (_db.Books.Find(id) == null)
+			throw new KeyNotFoundException($"There is no book with id {id}");
+		
 		var ee = _db.Books.Remove(_db.Books.Find(id));
 		_db.SaveChanges();
 
 		return ee.Entity;
 	}
+	
 }
